Hide the Internalizer_config class
